chore: _confirm function makefile

+++ makefile
@@ -4,7 +4,7 @@ REGION     ?=
 
 PLAN_ARGS = -var 'project_id=$(PROJECT_ID)' -var 'region=$(REGION)'
 
-.PHONY: activate_apis bootstrap_apply deploy destroy_gke core destroy_all help
+.PHONY: activate_apis bootstrap_apply deploy destroy_gke core destroy_all help confirm
 
 bootstrap: activate_apis bootstrap_apply
 
@@ -14,6 +14,7 @@ help: ## Display this help
 	@echo ""
 	@echo "Targets :"
 	@grep -E '^[a-zA-Z_-]+:.*?##' $(MAKEFILE_LIST) \
+		| grep -v '^_' \
 		| awk 'BEGIN {FS = ":.*?##"}; {printf "  \033[36m%-20s\033[0m %s\n", $$1, $$2}'
 
 activate_apis: ## Activate required APIs
@@ -49,6 +50,8 @@ deploy: ## Deploy infrastructure
 	terraform apply -var-file=../$(VARS_FILE)
 
 destroy_gke: ## Destroy GKE cluster and node pool
+	@echo "⚠️  You are about to destroy the GKE cluster and node pool. This action is irreversible."
+	@$(MAKE) _confirm
 	cd infra && \
 	terraform destroy -target=google_container_node_pool.primary_nodes -auto-approve && \
 	terraform destroy -target=google_container_cluster.gke_cluster -auto-approve
@@ -56,11 +59,7 @@ destroy_gke: ## Destroy GKE cluster and node pool
 
 destroy_all: ## Destroy all resources
 	@echo "⚠️  You are about to destroy all resources. This action is irreversible."
-	@read -p "❓ Are you sure you want to continue? (yes/no): " confirm; \
-	if [ "$$confirm" != "yes" ]; then \
-		echo "❌ Operation cancelled."; \
-		exit 1; \
-	fi
+	@$(MAKE) _confirm
 
 	cd infra && \
 	terraform init -backend-config="bucket=tfstate-$(PROJECT_ID)" -backend-config="prefix=gke" && \
@@ -86,3 +85,10 @@ destroy_all: ## Destroy all resources
 		compute.googleapis.com \
 		--project=$(PROJECT_ID) \
 		--force
+
+_confirm:
+	@read -p "❓ Are you sure you want to continue? (yes/no): " confirm; \
+	if [ "$$confirm" != "yes" ]; then \
+		echo "❌ Operation cancelled."; \
+		exit 1; \
+	fi
